# doc/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CalculEat API Documentation - OpenAPI 3.0",
    "description": "Sample description.",
    "version": "1.0.10"
  },
  "servers": [
    {
      "url": "http://hostname:7854/"
    }
  ],
...
  "paths": {
...
    "/api/1.0/user": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "List all of the users in DB.",
        "description": "Fetch all of the users from DB.",
        "operationId": "listUsers",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "The id of user that needs to be fetched",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "name": "role_id",
            "in": "query",
            "description": "The role_id (user type) that needs to be fetched",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "name": "first_name",
            "in": "query",
            "description": "The firstname that needs to be fetched",
            "required": false,
            "schema": {
              "type": "string",
              "example": "Yusuf"
            }
          },
          {
            "name": "last_name",
            "in": "query",
            "description": "The lastname that needs to be fetched",
            "required": false,
            "schema": {
              "type": "string",
              "example": "Kemal"
            }
          },
          {
            "name": "age",
            "in": "query",
            "description": "The age that needs to be fetched",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 23
            }
          },
          {
            "name": "gender",
            "in": "query",
            "description": "The gender that needs to be fetched",
            "required": false,
            "schema": {
              "type": "string",
              "example": "2000-06-23"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListUserResponseBody"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "read:users",
              "write:users"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Add a new user to the database",
        "description": "Add a new user to the database. All of the fields are not needed to create user.",
        "operationId": "createUser",
        "requestBody": {
          "description": "Create a new user in the DB.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseBody"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "405": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "write:users",
              "read:users"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Deletes the user from the database. This can only be done by the logged in user.",
        "description": "Deletes the specified user according to the JWT token provided in Authorization header.",
        "operationId": "deleteUser",
        "responses": {
          "200": {
            "description": "User successfuly deleted."
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "write:users",
              "read:users"
            ]
          }
        ]
      },
      "patch": {
        "tags": [
          "user"
        ],
        "summary": "Update user",
        "description": "This can only be done by the logged in user. **You can use only fields that you want to update.**",
        "operationId": "updateUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequestBody"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPATCHResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized request"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "write:users",
              "read:users"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserPATCHResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "format": "int64",
                "example": 1
              },
              "access_token": {
                "type": "string",
                "example": ""
              },
              "role_id": {
                "type": "integer",
                "format": "int64",
                "example": 1
              },
              "first_name": {
                "type": "string",
                "example": "John"
              },
              "last_name": {
                "type": "string",
                "example": "Doe"
              },
              "password": {
                "type": "string",
                "example": ""
              },
              "profile_img_filepath": {
                "type": "string",
                "example": "/images/profile_images/3/"
              },
              "gender": {
                "type": "string",
                "example": "M"
              },
              "born": {
                "type": "string",
                "example": "2000-06-23"
              },
              "age": {
                "type": "integer",
                "format": "int64",
                "example": 21
              },
              "email": {
                "type": "string",
                "example": "john@email.com"
              },
              "height": {
                "type": "integer",
                "format": "int64",
                "example": 182
              },
              "weight": {
                "type": "integer",
                "format": "int64",
                "example": 75
              },
              "goal": {
                "type": "string",
                "example": "Lose weight"
              },
              "sign_in_provider": {
                "type": "string",
                "example": "password"
              },
              "diseases": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "str1",
                  "str2"
                ]
              },
              "CreatedAt": {
                "type": "string",
                "example": "2023-08-20T17:03:43.442025+03:00"
              },
              "UpdatedAt": {
                "type": "string",
                "example": "2023-08-20T17:03:43.442025+03:00"
              },
              "deleted_on": {
                "type": "string",
                "example": ""
              }
            }
          },
          "msg": {
            "type": "string",
            "example": "Successful operation, object has been updated."
          }
        }
      },
      "ListUserResponseBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "access_token": {
            "type": "string",
            "example": ""
          },
          "role_id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "init_water_id": {
            "type": "integer",
            "format": "int64",
            "example": 0
          },
          "first_name": {
            "type": "string",
            "example": "John"
          },
          "last_name": {
            "type": "string",
            "example": "Doe"
          },
          "password": {
            "type": "string",
            "example": ""
          },
          "profile_img_filepath": {
            "type": "string",
            "example": ""
          },
          "gender": {
            "type": "string",
            "example": "M"
          },
          "born": {
            "type": "string",
            "example": ""
          },
          "age": {
            "type": "integer",
            "format": "int64",
            "example": 21
          },
          "email": {
            "type": "string",
            "example": ""
          },
          "height": {
            "type": "integer",
            "format": "int64",
            "example": 0
          },
          "weight": {
            "type": "integer",
            "format": "int64",
            "example": 0
          },
          "goal": {
            "type": "string",
            "example": ""
          },
          "sign_in_provider": {
            "type": "string",
            "example": ""
          },
          "diseases": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              null
            ]
          },
          "CreatedAt": {
            "type": "string",
            "example": "0000-00-00T00:00:00.000000+03:00"
          },
          "UpdatedAt": {
            "type": "string",
            "example": "0000-00-00T00:00:00.000000+03:00"
          },
          "deleted_on": {
            "type": "string",
            "example": ""
          }
        }
      },
      "UserResponseBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "access_token": {
            "type": "string",
            "example": ""
          },
          "role_id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "init_water_id": {
            "type": "integer",
            "format": "int64",
            "example": 2
          },
          "first_name": {
            "type": "string",
            "example": "John"
          },
          "last_name": {
            "type": "string",
            "example": "Doe"
          },
          "password": {
            "type": "string",
            "example": ""
          },
          "profile_img_filepath": {
            "type": "string",
            "example": "/images/profile_images/3/"
          },
          "gender": {
            "type": "string",
            "example": "M"
          },
          "born": {
            "type": "string",
            "example": "2000-06-23"
          },
          "age": {
            "type": "integer",
            "format": "int64",
            "example": 21
          },
          "email": {
            "type": "string",
            "example": "john@email.com"
          },
          "height": {
            "type": "integer",
            "format": "int64",
            "example": 182
          },
          "weight": {
            "type": "integer",
            "format": "int64",
            "example": 75
          },
          "goal": {
            "type": "string",
            "example": "Lose weight"
          },
          "sign_in_provider": {
            "type": "string",
            "example": "password"
          },
          "diseases": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "str1",
              "str2"
            ]
          },
          "CreatedAt": {
            "type": "string",
            "example": "2023-08-20T17:03:43.442025+03:00"
          },
          "UpdatedAt": {
            "type": "string",
            "example": "2023-08-20T17:03:43.442025+03:00"
          },
          "deleted_on": {
            "type": "string",
            "example": ""
          }
        }
      },
...
      "UserRequestBody": {
        "type": "object",
        "properties": {
          "role_id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "first_name": {
            "type": "string",
            "example": "John"
          },
          "last_name": {
            "type": "string",
            "example": "Doe"
          },
          "password": {
            "type": "string",
            "example": "SuperSecr3t"
          },
          "profile_img_filepath": {
            "type": "string",
            "example": "/images/profile_images/3/"
          },
          "gender": {
            "type": "string",
            "example": "M"
          },
          "born": {
            "type": "string",
            "example": "2000-06-23"
          },
          "age": {
            "type": "integer",
            "format": "int64",
            "example": 21
          },
          "email": {
            "type": "string",
            "example": "john@email.com"
          },
          "height": {
            "type": "integer",
            "format": "int64",
            "example": 182
          },
          "weight": {
            "type": "integer",
            "format": "int64",
            "example": 75
          },
          "goal": {
            "type": "string",
            "example": "Lose weight"
          },
          "diseases": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "str1",
              "str2"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}